requirements.txt
b2sdk
python-dotenv